renamed files for js convention
