+++ Frontend/src/register/pages/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
+import RegistrationList from '../components/RegistrationList'
+import RegistrationPanel from '../components/RegistrationPanel'
 import styles from './index.module.scss'
-import RegistrationList from '../components/registration_list'
-import RegistrationPanel from '../components/registration_panel'
 
 export default function App() {
   return (
